@@ -5,13 +5,12 @@
     "type": "symfony-bundle",
     "license": "MIT",
     "require": {
-        "php": "^5.5|^7.0",
-        "contentful/contentful": "^2.0",
-        "symfony/framework-bundle": "^2.7|^3.0",
-        "symfony/lts": "^2.0|^3.0"
+        "php": "^7.0",
+        "contentful/contentful": "^3.3",
+        "symfony/framework-bundle": "^3.4|^4.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8|^5.7"
+        "phpunit/phpunit": "^6.0|^7.0"
     },
     "autoload": {
         "psr-4": {
@@ -25,7 +24,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.1.0-dev"
+            "dev-master": "3.0.0-dev"
         }
     }
 }
